fix(pipe): use u64 for sample counter to prevent overflow
u32 overflows after ~49 days at 1 sample/sec. u64 is effectively
unlimited. Fixes spec requirement for pipe mode counter safety.

+++ src/cli.rs
@@ -40,7 +40,7 @@ pub enum Command {
     Pipe {
         /// Number of samples (0 = infinite)
         #[arg(short, long, default_value_t = 0)]
-        samples: u32,
+        samples: u64,
     },
 
     /// Start HTTP API server
